requirements.txt
tensorflow==2.4.0
streamlit==0.73.0
Pillow==8.0.1
